Refactor to ideal step architecture
